+++ android/app/src/main/java/org/openbmap/events/onSessionStart.java
@@ -21,7 +21,7 @@ package org.openbmap.events;
 import org.openbmap.Constants;
 
 public class onSessionStart {
-    public final int session;
+    public final long session;
 
     /**
      * Default constructor: no session id provided, database will auto-assign session id
@@ -34,7 +34,7 @@ public class onSessionStart {
      * Constructor to resume an existing session
      * @param session session id to resume
      */
-    public onSessionStart(int session) {
+    public onSessionStart(long session) {
         this.session = session;
     }
